// src/interfaces/enum.ts
export enum ControllerPrefix {
  user = 'user',
}
export enum RegisterType {
  PHONE = 'phone',
  EMAIL = 'email',
}
export enum LoginType {
  PHONE = 'phone',
  EMAIL = 'email',
  verify = 'verifyCode',
}
export enum CustomCode {
  SQLERROR = 209,
}
export enum RedisKeyPrefix {
  user = 'user',
}
export enum UserStatus {
  BANNED = 0,
  ACCESS = 1,
}