// src/interfaces/interface.ts
export interface IUser {
  id: string;
  phone: string;
  avatar: string;
  nickname: string;
  createTime: string;
  password: string;
  IDCard: string;
  birthday: string;
  sex: number;
  status: number;
  isReceivedRequest: number;
  email: string;
  username: string;
  updateTime: string;
  token: string;
}